fix(subprocess): address copilot review round 12 — usize count fields in both OSC trackers
Two threads, both about the same code-quality issue:

1. **proxy.zig:876** + **osc133.zig:238**: `count` fields in
   `Osc7` and `edge_count` in `Osc133` were `u8`. The proxy's
   merge loop compares them against `usize` cursors
   (`ci < osc7_tracker.count`) and slices `edges[0..edge_count]`
   — Zig 0.16 implicitly coerces the u8 to usize so the code
   compiles, but the implicit mix is a future-footgun and adds
   visual noise at every comparison site.

   Fix: changed both counts to `usize`. The 16/32-entry caps are
   far below `u8` range, so the storage cost is irrelevant.
   Tests updated to use `@as(usize, …)` for the new field type
   (sed-driven, 16 sites in osc7.zig).

+++ src/osc133.zig
@@ -61,7 +61,11 @@ pub const Osc133 = struct {
     /// the wrong frame when `OSC 7` and `;C` co-occur in a single
     /// read chunk.
     edge_offsets: [32]u32 = undefined,
-    edge_count: u8 = 0,
+    /// `usize` rather than `u8` despite the bounded array size so
+    /// `edges[0..edge_count]` slicing and `edgeOffset(idx)` indexing
+    /// don't need explicit casts at every call site. The 32-entry
+    /// ring caps the value far below `u8` range anyway.
+    edge_count: usize = 0,
     /// Byte index within the current `feed()` call. Stamped onto
     /// each edge in `edge_offsets`. Reset at the start of every
     /// `feed()` invocation; the proxy reads it indirectly via the
